refactor(test): rename directories
