+++ packages/integration-tests/src/tests/console/rbac.test.ts
@@ -4,7 +4,7 @@ import {
   expectToClickDetailsPageOption,
   goToAdminConsole,
   waitForToast,
-} from '#src/ui-helpers/index.js';
+} from '#src/page-helpers/index.js';
 import {
   expectNavigation,
   appendPathname,
