+++ packages/integration-tests/src/tests/console/sign-in-experience/helpers.ts
@@ -4,7 +4,7 @@ import {
   expectToSaveChanges,
   expectConfirmModalAndAct,
   waitForToast,
-} from '#src/ui-helpers/index.js';
+} from '#src/page-helpers/index.js';
 
 export const waitForFormCard = async (page: Page, title: string) => {
   await expect(page).toMatchElement('div[class$=tabContent] div[class$=card] div[class$=title]', {
