+++ packages/integration-tests/src/tests/console/sign-in-experience/sign-in-preview.test.ts
@@ -2,7 +2,7 @@ import { type Nullable } from '@silverhand/essentials';
 import { type Page, type Target } from 'puppeteer';
 
 import { logtoConsoleUrl as logtoConsoleUrlString, logtoUrl } from '#src/constants.js';
-import { goToAdminConsole } from '#src/ui-helpers/index.js';
+import { goToAdminConsole } from '#src/page-helpers/index.js';
 import { expectNavigation, appendPathname } from '#src/utils.js';
 
 import { expectToSelectPreviewLanguage, waitForFormCard } from './helpers.js';
