+++ packages/integration-tests/src/tests/console/sign-in-experience/sign-up-and-sign-in/connector-setup-helpers.ts
@@ -2,7 +2,7 @@ import { ConnectorType } from '@logto/schemas';
 import { type Page } from 'puppeteer';
 
 import { logtoConsoleUrl as logtoConsoleUrlString } from '#src/constants.js';
-import { expectToClickDetailsPageOption, waitForToast } from '#src/ui-helpers/index.js';
+import { expectToClickDetailsPageOption, waitForToast } from '#src/page-helpers/index.js';
 import { expectNavigation, appendPathname } from '#src/utils.js';
 
 import {
