+++ packages/integration-tests/src/tests/console/sign-in-experience/sign-up-and-sign-in/helpers.ts
