+++ packages/integration-tests/src/tests/console/sign-in-experience/sign-up-and-sign-in/sad-path.test.ts
@@ -3,7 +3,7 @@ import {
   expectToClickNavTab,
   goToAdminConsole,
   expectToSaveChanges,
-} from '#src/ui-helpers/index.js';
+} from '#src/page-helpers/index.js';
 import { expectNavigation, appendPathname } from '#src/utils.js';
 
 import { expectToSaveSignInExperience, waitForFormCard } from '../helpers.js';
