+++ packages/integration-tests/src/tests/main-flow/bootstrap.test.ts
@@ -1,0 +1,11 @@
+/**
+ * NOTE: This test suite assumes test cases will run sequentially (which is Jest default).
+ * Parallel execution will lead to errors.
+ */
+// Tip: See https://github.com/argos-ci/jest-puppeteer/blob/main/packages/expect-puppeteer/README.md
+// for convenient expect methods
+describe('smoke testing for the main flow', () => {
+  it('should not explode', async () => {
+    expect(true);
+  });
+});
